requirements.txt
requests>=2.32.5,<3.0.0
pytz>=2025.2,<2026.0
PyYAML>=6.0.3,<7.0.0
websockets>=13.0,<14.0
boto3>=1.35.0,<2.0.0
feedparser>=6.0.0,<7.0.0
beautifulsoup4>=4.12.3,<5.0.0
google-generativeai>=0.5.0,<1.0.0
fastapi>=0.111.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
python-multipart>=0.0.9,<0.1.0
jieba>=0.42.1,<1.0.0
scikit-learn>=1.3.0,<2.0.0
numpy>=1.24.0,<2.0.0

# 国内AI服务提供商SDK（可选，根据需要安装）
# 百度文心一言SDK
# erniebot-agent>=0.1.0,<1.0.0
# 
# 阿里通义千问SDK
# dashscope>=1.0.0,<2.0.0
